+2 tests compile ; warning hunt

+++ trunk/linbox/tests/test-dense.C
@@ -7,20 +7,20 @@
  *
  * --------------------------------------------------------
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -179,7 +179,7 @@ static bool testVandermonde (Field &F, size_t n, int iterations = 1, int N = 1)
 		commentator().start (buf);
 
 		/* Evaluation points */
-		for (j = 0; j < n; j++) {
+		for (j = 0; j < (int) n; j++) {
 			bool flag = true;
 
 			// Make sure points are all distinct
@@ -197,16 +197,16 @@ static bool testVandermonde (Field &F, size_t n, int iterations = 1, int N = 1)
 		printVector<Field> (F, report, x);
 
 		/* Build the Vandermonde matrix */
-		for (j = 0; j < n; j++) {
+		for (j = 0; j < (int) n; j++) {
 			F.init (t, 1);
 
-			for (k = 0; k < n; k++) {
+			for (k = 0; k < (int) n; k++) {
 				V.setEntry (j, k, t);
 				F.mulin (t, x[j]);
 			}
 		}
 
-		for (j = 0; j < N; j++) {
+		for (j = 0; j < (int) N; j++) {
 			inner_iter_passed = true;
 
 			/* Random vector of evaluation results */
@@ -228,7 +228,7 @@ static bool testVandermonde (Field &F, size_t n, int iterations = 1, int N = 1)
 			report << "Interpolation results: ";
 			printVector<Field> (F, report, f);
 
-			for (k = 0; k < n; k++)
+			for (k = 0; k < (int) n; k++)
 				if (!F.areEqual (f[k], v[k]))
 					ret = inner_iter_passed = false;
